requirements.txt
# Core LangChain and LangGraph for orchestration
langchain
langgraph
langchain-openai
langchain-google-genai

# Additional dependencies for LangGraph
pydantic
pydantic-settings

# Environment and configuration
python-dotenv>=1.0.0

# Web API framework
fastapi
uvicorn[standard]

# HTTP requests
requests

# Git operations
GitPython

# AST parsing
tree-sitter>=0.20.1
tree-sitter-python>=0.20.1
tree-sitter-java>=0.20.1
tree-sitter-kotlin>=0.20.1
tree-sitter-xml==0.7.0
tree-sitter-javascript>=0.20.1
tree-sitter-language-pack>=0.7.3

# Code complexity and quality metrics
radon

# RAG and Vector Database
qdrant-client
sentence-transformers
transformers
torch
llama-index
llama-index-core
llama-index-embeddings-huggingface

# Testing framework
pytest
pytest-cov
pytest-asyncio
pytest-benchmark
freezegun

# HTTP client for testing FastAPI
httpx 

# Task queue and async processing
celery
redis
aiofiles

# Neo4j dependency
neo4j>=5.14.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Authentication and security
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0